requirements.txt
Django==2.2.6
djangorestframework==3.10.3
flake8==3.5.0
flake8-commas==2.0.0
flake8-quotes==1.0.0
